sghss-api/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
pydantic==2.5.0